Slow build.rs / rust analyzer #913

+++ server/build.rs
@@ -19,7 +19,7 @@ struct Dirs {
 
 fn main() -> std::io::Result<()> {
     // Uncomment this line if you want faster builds during development
-    return Ok(());
+    // return Ok(());
     const BROWSER_ROOT: &str = "../browser/";
     let dirs: Dirs = {
         Dirs {
@@ -31,30 +31,49 @@ fn main() -> std::io::Result<()> {
     };
     println!("cargo:rerun-if-changed={}", BROWSER_ROOT);
 
-    if should_build(&dirs) {
-        build_js(&dirs);
-        let _ = fs::remove_dir_all(&dirs.js_dist_tmp);
-        dircpy::copy_dir(&dirs.js_dist_source, &dirs.js_dist_tmp)?;
-    } else if dirs.js_dist_tmp.exists() {
-        p!("Found {}, skipping copy", dirs.js_dist_tmp.display());
-    } else {
-        p!(
-            "Could not find {} , copying from {}",
-            dirs.js_dist_tmp.display(),
-            dirs.js_dist_source.display()
-        );
-        dircpy::copy_dir(&dirs.js_dist_source, &dirs.js_dist_tmp)?;
-    }
+    let is_check_like = profile == "debug" && opt_level == "0";
 
-    // Makes the static files available for compilation
-    static_files::resource_dir(&dirs.js_dist_tmp)
-        .build()
-        .unwrap_or_else(|_e| {
-            panic!(
-                "failed to open data browser assets from {}",
-                dirs.js_dist_tmp.display()
-            )
-        });
+    if is_check_like {
+        println!("cargo:rerun-if-changed=build.rs");
+        // Skip the heavy logic
+        println!("Skipping build.rs logic for cargo check/clippy.");
+    } else {
+        const BROWSER_ROOT: &str = "../browser/";
+        println!("cargo:rerun-if-changed={}", BROWSER_ROOT);
+        let dirs: Dirs = {
+            Dirs {
+                js_dist_source: PathBuf::from("../browser/data-browser/dist"),
+                js_dist_tmp: PathBuf::from("./assets_tmp"),
+                src_browser: PathBuf::from("../browser/data-browser/src"),
+                browser_root: PathBuf::from(BROWSER_ROOT),
+            }
+        };
+        //
+        if should_build(&dirs) {
+            build_js(&dirs);
+            let _ = fs::remove_dir_all(&dirs.js_dist_tmp);
+            dircpy::copy_dir(&dirs.js_dist_source, &dirs.js_dist_tmp)?;
+        } else if dirs.js_dist_tmp.exists() {
+            p!("Found {}, skipping copy", dirs.js_dist_tmp.display());
+        } else {
+            p!(
+                "Could not find {} , copying from {}",
+                dirs.js_dist_tmp.display(),
+                dirs.js_dist_source.display()
+            );
+            dircpy::copy_dir(&dirs.js_dist_source, &dirs.js_dist_tmp)?;
+        }
+
+        // Makes the static files available for compilation
+        static_files::resource_dir(&dirs.js_dist_tmp)
+            .build()
+            .unwrap_or_else(|_e| {
+                panic!(
+                    "failed to open data browser assets from {}",
+                    dirs.js_dist_tmp.display()
+                )
+            });
+    }
 
     Ok(())
 }
